+++ src/core.sh
@@ -25,11 +25,11 @@
 #
 # CORE
 # C : 2021/03/13
-# M : 2021/03/16
+# M : 2021/03/18
 # D : Utility functions
 
 # shellcheck disable=SC2034
-__version="0.6.2"
+__version="0.7"
 
 CONFIG="$HOME/.config/gmi/config"